requirements.txt
fastapi
uvicorn
onnx
onnxruntime
torch
torchvision
numpy
Pillow